+++ changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.1.0] - 2026-02-18
+
+### Added
+- `setPinMode(uint8_t mask)` - nastaveni smeru pinu maskou (1=input, 0=output)
+- `setPinMode(uint8_t pin, bool input)` - nastaveni smeru jednoho pinu
+- Vstupni piny se pri kazdem `writeAll()` a `writePin()` automaticky drzi na HIGH
+
 ## [1.0.0] - 2026-02-13
 
 ### Added